+++ static/app/views/performance/mobile/ui/components/charts/topScreensChart.tsx
